make fabric code path compile again
