+++ projects/rocshmem/src/memory/hip_allocator_vmm_fabric.cpp
@@ -22,7 +22,7 @@
  * IN THE SOFTWARE.
  *****************************************************************************/
 
-#include "hip_allocator_vmm_fabric.hpp"
+#include "hip_allocator.hpp"
 #include "hip_allocator_vmm_common.hpp"
 
 #if HIP_VERSION >= 70000000
